refactor: relocate test module
